Re-implement data collection via rotary encoder
Changed arduino code so that 1) the arduino actually sends Matlab the data collected by the rotary encoder and 2) the arduino can collect the same amount of encoder data as camera data.

Modified Matlab code to query the arduino for the correct amount of data and then actually save the encoder date. Note that I had to increase the input buffer size in Launch.m in order to be able to collect data for up to 2 seconds.