Fixing bug where arduino fails to start long trials
For some reason, Neuroblinks could run a single trial of 5200 ms or greater duration, but it could not run a second trial. The camera would be triggered, but then the trial would stop before any stimuli could be triggered. I believe this bug had to do with the update function in statemachine.cpp. In the previous version of the code, the camera, etc. would be initialized to have num_repeats=1. Then, num_repeats would be decremented to 0 in update during the first trial. This value would never be reset to 1, and it would instead be decremented progressively more negative with subsequent trials. This could cause the state machines to get stuck in INTERPULSE and never actually turn off. When I added lines to refresh the number of repeats for camera and US stimuli to 1, I could run multiple trials of >=5.2 s duration.

I also increased the maximum array size for the encoder readings, so that the arduino will be able to store up to 15s of encoder data instead of 5s of encoder data at a 5 ms inter-sample interval.